app.py refactoring
