+++ app.py
@@ -506,336 +506,6 @@ def index():
     """
     return render_template('index.html')
 
-@app.route('/api/analysis/start', methods=['POST'])
-def start_analysis_endpoint():
-    """
-    Start the music analysis process for recent albums.
-    This endpoint enqueues a main analysis task.
-    Note: Starting a new analysis task will archive previously successful tasks by setting their status to REVOKED.
-    ---
-    tags:
-      - Analysis
-    requestBody:
-      description: Configuration for the analysis task.
-      required: false
-      content:
-        application/json:
-          schema:
-            type: object
-            properties:
-              jellyfin_url:
-                type: string
-                description: URL of the Jellyfin server.
-                default: "Configured JELLYFIN_URL"
-              jellyfin_user_id:
-                type: string
-                description: Jellyfin User ID.
-                default: "Configured JELLYFIN_USER_ID"
-              jellyfin_token:
-                type: string
-                description: Jellyfin API Token.
-                default: "Configured JELLYFIN_TOKEN"
-              num_recent_albums:
-                type: integer
-                description: Number of recent albums to process.
-                default: "Configured NUM_RECENT_ALBUMS"
-              top_n_moods:
-                type: integer
-                description: Number of top moods to extract per track.
-                default: "Configured TOP_N_MOODS"
-    responses:
-      202:
-        description: Analysis task successfully enqueued.
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                task_id:
-                  type: string
-                  description: The ID of the enqueued main analysis task.
-                task_type:
-                  type: string
-                  description: Type of the task (e.g., main_analysis).
-                  example: main_analysis
-                status:
-                  type: string
-                  description: The initial status of the job in the queue (e.g., queued).
-      400:
-        description: Invalid input.
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                error:
-                  type: string
-      500:
-        description: Server error during task enqueue.
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                error:
-                  type: string
-                message:
-                  type: string
-    """
-    data = request.json or {}
-    jellyfin_url = data.get('jellyfin_url', JELLYFIN_URL)
-    jellyfin_user_id = data.get('jellyfin_user_id', JELLYFIN_USER_ID)
-    jellyfin_token = data.get('jellyfin_token', JELLYFIN_TOKEN)
-    num_recent_albums = int(data.get('num_recent_albums', NUM_RECENT_ALBUMS))
-    top_n_moods = int(data.get('top_n_moods', TOP_N_MOODS))
-
-    job_id = str(uuid.uuid4())
-
-    # Clean up details of previously successful tasks before starting a new one
-    clean_successful_task_details_on_new_start()
-    save_task_status(job_id, "main_analysis", TASK_STATUS_PENDING, details={"message": "Task enqueued."})
-
-    # Enqueue task using a string path to its function.
-    # This completely breaks the circular import cycle between app.py and the tasks module.
-    job = rq_queue_high.enqueue(
-        'tasks.analysis.run_analysis_task',
-        args=(jellyfin_url, jellyfin_user_id, jellyfin_token, num_recent_albums, top_n_moods),
-        job_id=job_id,
-        description="Main Music Analysis",
-        retry=Retry(max=3),
-        job_timeout=-1 # No timeout
-    )
-    return jsonify({"task_id": job.id, "task_type": "main_analysis", "status": job.get_status()}), 202
-
-@app.route('/api/clustering/start', methods=['POST'])
-def start_clustering_endpoint():
-    """
-    Start the music clustering and playlist generation process.
-    This endpoint enqueues a main clustering task.
-    Note: Starting a new clustering task will archive previously successful tasks by setting their status to REVOKED.
-    ---
-    tags:
-      - Clustering
-    requestBody:
-      description: Configuration for the clustering task.
-      required: true
-      content:
-        application/json:
-          schema:
-            type: object
-            properties:
-              clustering_method:
-                type: string
-                description: Algorithm to use for clustering (e.g., kmeans, dbscan, gmm).
-                default: "Configured CLUSTER_ALGORITHM"
-              num_clusters_min:
-                type: integer
-                description: Minimum number of clusters (for kmeans/gmm).
-                default: "Configured NUM_CLUSTERS_MIN"
-              num_clusters_max:
-                type: integer
-                description: Maximum number of clusters (for kmeans/gmm).
-                default: "Configured NUM_CLUSTERS_MAX"
-              dbscan_eps_min:
-                type: number
-                format: float
-                description: Minimum epsilon for DBSCAN.
-                default: "Configured DBSCAN_EPS_MIN"
-              dbscan_eps_max:
-                type: number
-                format: float
-                description: Maximum epsilon for DBSCAN.
-                default: "Configured DBSCAN_EPS_MAX"
-              dbscan_min_samples_min:
-                type: integer
-                description: Minimum min_samples for DBSCAN.
-                default: "Configured DBSCAN_MIN_SAMPLES_MIN"
-              dbscan_min_samples_max:
-                type: integer
-                description: Maximum min_samples for DBSCAN.
-                default: "Configured DBSCAN_MIN_SAMPLES_MAX"
-              gmm_n_components_min:
-                type: integer
-                description: Minimum number of components for GMM.
-                default: "Configured GMM_N_COMPONENTS_MIN"
-              gmm_n_components_max:
-                type: integer
-                description: Maximum number of components for GMM.
-                default: "Configured GMM_N_COMPONENTS_MAX"
-              pca_components_min:
-                type: integer
-                description: Minimum number of PCA components.
-                default: "Configured PCA_COMPONENTS_MIN"
-              pca_components_max:
-                type: integer
-                description: Maximum number of PCA components.
-                default: "Configured PCA_COMPONENTS_MAX"
-              clustering_runs:
-                type: integer
-                description: Number of clustering iterations to perform.
-                default: "Configured CLUSTERING_RUNS"
-              score_weight_diversity:
-                type: number
-                format: float
-                description: Weight for the inter-playlist mood diversity score component.
-                default: "Configured SCORE_WEIGHT_DIVERSITY"
-              score_weight_silhouette:
-                type: number
-                format: float
-                description: Weight for the Silhouette score component.
-                default: "Configured SCORE_WEIGHT_SILHOUETTE"
-              score_weight_davies_bouldin:
-                type: number
-                format: float
-                description: Weight for the Davies-Bouldin score component (higher is better for score calculation).
-                default: "Configured SCORE_WEIGHT_DAVIES_BOULDIN"
-              score_weight_calinski_harabasz:
-                type: number
-                format: float
-                description: Weight for the Calinski-Harabasz score component (higher is better).
-                default: "Configured SCORE_WEIGHT_CALINSKI_HARABASZ"
-              score_weight_purity:
-                type: number
-                format: float
-                description: Weight for playlist purity (intra-playlist mood consistency).
-                default: "Configured SCORE_WEIGHT_PURITY"
-              score_weight_other_feature_diversity:
-                type: number
-                format: float
-                description: Weight for inter-playlist diversity of other features (e.g., danceability).
-                default: "Configured SCORE_WEIGHT_OTHER_FEATURE_DIVERSITY"
-              score_weight_other_feature_purity:
-                type: number
-                format: float
-                description: Weight for intra-playlist consistency of other features (e.g., danceability).
-                default: "Configured SCORE_WEIGHT_OTHER_FEATURE_PURITY"
-              min_songs_per_genre_for_stratification:
-                type: integer
-                description: Minimum number of songs to target per stratified genre.
-                default: "Configured MIN_SONGS_PER_GENRE_FOR_STRATIFICATION"
-              stratified_sampling_target_percentile:
-                type: integer
-                description: Percentile of genre song counts to use for target songs per genre.
-                minimum: 0
-                maximum: 100
-                default: "Configured STRATIFIED_SAMPLING_TARGET_PERCENTILE"
-              max_songs_per_cluster:
-                type: integer
-                description: Maximum number of songs per generated playlist/cluster.
-                default: "Configured MAX_SONGS_PER_CLUSTER"
-              ai_model_provider:
-                type: string
-                description: AI provider for playlist naming (OLLAMA, GEMINI, NONE).
-                default: "Configured AI_MODEL_PROVIDER"
-              ollama_server_url:
-                type: string
-                description: Override for the Ollama server URL for this run.
-                nullable: true
-                default: "Defaults to server-configured OLLAMA_SERVER_URL"
-              ollama_model_name:
-                type: string
-                description: Override for the Ollama model name for this run.
-                nullable: true
-                default: "Defaults to server-configured OLLAMA_MODEL_NAME"
-              gemini_api_key:
-                type: string
-                description: Override for the Gemini API key for this run.
-                nullable: true
-                default: "Defaults to server-configured GEMINI_API_KEY"
-              gemini_model_name:
-                type: string
-                description: Override for the Gemini model name for this run.
-                nullable: true
-                default: "Defaults to server-configured GEMINI_MODEL_NAME"
-              top_n_moods:
-                type: integer
-                description: Number of top moods to consider for clustering feature vectors (uses the first N from global MOOD_LABELS).
-                default: "Configured TOP_N_MOODS"
-              enable_clustering_embeddings:
-                type: boolean
-                description: Whether to use embeddings for clustering (True) or score_vector (False).
-                default: false
-    responses:
-      202:
-        description: Clustering task successfully enqueued.
-        content:
-          application/json:
-            schema:
-              type: object
-              properties:
-                task_id:
-                  type: string
-                  description: The ID of the enqueued main clustering task.
-                task_type:
-                  type: string
-                  description: Type of the task (e.g., main_clustering).
-                  example: main_clustering
-                status:
-                  type: string
-                  description: The initial status of the job in the queue (e.g., queued).
-      409:
-        description: An active clustering task is already in progress.
-        content:
-            application/json:
-                schema:
-                    type: object
-                    properties:
-                        error:
-                            type: string
-                        task_id:
-                            type: string
-                        status:
-                            type: string
-    """
-    # Enqueue by string to avoid circular imports
-    data = request.json
-    job_id = str(uuid.uuid4())
-
-    # Clean up details of previously successful tasks before starting a new one
-    clean_successful_task_details_on_new_start()
-    save_task_status(job_id, "main_clustering", TASK_STATUS_PENDING, details={"message": "Task enqueued."})
-
-    job = rq_queue_high.enqueue(
-        'tasks.clustering.run_clustering_task', # Enqueue by string path
-        kwargs={ # Pass all arguments as a dictionary
-            "clustering_method": data.get('clustering_method', CLUSTER_ALGORITHM),
-            "num_clusters_min": int(data.get('num_clusters_min', NUM_CLUSTERS_MIN)),
-            "num_clusters_max": int(data.get('num_clusters_max', NUM_CLUSTERS_MAX)),
-            "dbscan_eps_min": float(data.get('dbscan_eps_min', DBSCAN_EPS_MIN)),
-            "dbscan_eps_max": float(data.get('dbscan_eps_max', DBSCAN_EPS_MAX)),
-            "dbscan_min_samples_min": int(data.get('dbscan_min_samples_min', DBSCAN_MIN_SAMPLES_MIN)),
-            "dbscan_min_samples_max": int(data.get('dbscan_min_samples_max', DBSCAN_MIN_SAMPLES_MAX)),
-            "gmm_n_components_min": int(data.get('gmm_n_components_min', GMM_N_COMPONENTS_MIN)),
-            "gmm_n_components_max": int(data.get('gmm_n_components_max', GMM_N_COMPONENTS_MAX)),
-            "pca_components_min": int(data.get('pca_components_min', PCA_COMPONENTS_MIN)),
-            "pca_components_max": int(data.get('pca_components_max', PCA_COMPONENTS_MAX)),
-            "num_clustering_runs": int(data.get('clustering_runs', CLUSTERING_RUNS)),
-            "max_songs_per_cluster": int(data.get('max_songs_per_cluster', MAX_SONGS_PER_CLUSTER)),
-            "min_songs_per_genre_for_stratification": int(data.get('min_songs_per_genre_for_stratification', MIN_SONGS_PER_GENRE_FOR_STRATIFICATION)),
-            "stratified_sampling_target_percentile": int(data.get('stratified_sampling_target_percentile', STRATIFIED_SAMPLING_TARGET_PERCENTILE)),
-            "score_weight_diversity": float(data.get('score_weight_diversity', SCORE_WEIGHT_DIVERSITY)),
-            "score_weight_silhouette": float(data.get('score_weight_silhouette', SCORE_WEIGHT_SILHOUETTE)),
-            "score_weight_davies_bouldin": float(data.get('score_weight_davies_bouldin', SCORE_WEIGHT_DAVIES_BOULDIN)),
-            "score_weight_calinski_harabasz": float(data.get('score_weight_calinski_harabasz', SCORE_WEIGHT_CALINSKI_HARABASZ)),
-            "score_weight_purity": float(data.get('score_weight_purity', SCORE_WEIGHT_PURITY)),
-            "score_weight_other_feature_diversity": float(data.get('score_weight_other_feature_diversity', SCORE_WEIGHT_OTHER_FEATURE_DIVERSITY)),
-            "score_weight_other_feature_purity": float(data.get('score_weight_other_feature_purity', SCORE_WEIGHT_OTHER_FEATURE_PURITY)),
-            "ai_model_provider": data.get('ai_model_provider', AI_MODEL_PROVIDER).upper(),
-            "ollama_server_url": data.get('ollama_server_url', OLLAMA_SERVER_URL),
-            "ollama_model_name": data.get('ollama_model_name', OLLAMA_MODEL_NAME),
-            "gemini_api_key": data.get('gemini_api_key', GEMINI_API_KEY),
-            "gemini_model_name": data.get('gemini_model_name', GEMINI_MODEL_NAME),
-            "top_n_moods_for_clustering": int(data.get('top_n_moods', TOP_N_MOODS)),
-            "enable_clustering_embeddings": data.get('enable_clustering_embeddings', ENABLE_CLUSTERING_EMBEDDINGS),
-        },
-        job_id=job_id,
-        description="Main Music Clustering",
-        retry=Retry(max=3),
-        job_timeout=-1 # No timeout
-    )
-    return jsonify({"task_id": job.id, "task_type": "main_clustering", "status": job.get_status()}), 202
-
-
 @app.route('/api/status/<task_id>', methods=['GET'])
 def get_task_status_endpoint(task_id):
     """
@@ -1423,93 +1093,6 @@ def get_playlists_endpoint():
         playlists_data[row['playlist_name']].append({"item_id": row['item_id'], "title": row['title'], "author": row['author']})
     return jsonify(dict(playlists_data)), 200
 
-@app.route('/api/similar_tracks/<item_id>', methods=['GET'])
-def get_similar_tracks_endpoint(item_id):
-    """
-    Find similar tracks for a given track ID.
-    This endpoint uses the pre-loaded Annoy index for fast lookups.
-    ---
-    tags:
-      - Similarity
-    parameters:
-      - name: item_id
-        in: path
-        required: true
-        description: The Jellyfin Item ID of the track to find neighbors for.
-        schema:
-          type: string
-      - name: n
-        in: query
-        description: The number of similar tracks to return.
-        schema:
-          type: integer
-          default: 10
-    responses:
-      200:
-        description: A list of similar tracks with their details.
-        content:
-          application/json:
-            schema:
-              type: array
-              items:
-                $ref: '#/components/schemas/SimilarTrack'
-      404:
-        description: Target track not found in the index or no similar tracks found.
-      500:
-        description: Server error, e.g., Annoy index not loaded.
-    components:
-      schemas:
-        SimilarTrack:
-          type: object
-          properties:
-            item_id:
-              type: string
-            title:
-              type: string
-            author:
-              type: string
-            distance:
-              type: number
-              format: float
-              description: The similarity distance (lower is more similar).
-    """
-    # Import locally to prevent circular dependency at module load time.
-    from tasks.annoy_manager import find_nearest_neighbors_by_id
-    num_neighbors = request.args.get('n', 10, type=int)
-    try:
-        # This function now returns a list of dicts with {'item_id': id, 'distance': dist}
-        neighbor_results = find_nearest_neighbors_by_id(item_id, n=num_neighbors)
-        if not neighbor_results:
-            return jsonify({"error": "Target track not found in index or no neighbors found."}), 404
-
-        # Get the full track details for the neighbors
-        neighbor_ids = [n['item_id'] for n in neighbor_results]
-        neighbor_details = get_score_data_by_ids(neighbor_ids)
-
-        # Create a map for easy lookup
-        details_map = {d['item_id']: d for d in neighbor_details}
-        distance_map = {n['item_id']: n['distance'] for n in neighbor_results}
-
-        # Combine details with distance, preserving the order from Annoy
-        final_results = []
-        for neighbor_id in neighbor_ids:
-            if neighbor_id in details_map:
-                track_info = details_map[neighbor_id]
-                final_results.append({
-                    "item_id": track_info['item_id'],
-                    "title": track_info['title'],
-                    "author": track_info['author'],
-                    "distance": distance_map[neighbor_id]
-                })
-
-        return jsonify(final_results)
-    except RuntimeError as e:
-        logger.error(f"Runtime error finding neighbors for {item_id}: {e}")
-        return jsonify({"error": str(e)}), 500
-    except Exception as e:
-        logger.error(f"Unexpected error finding neighbors for {item_id}: {e}", exc_info=True)
-        return jsonify({"error": "An unexpected error occurred."}), 500
-
 def listen_for_index_reloads():
     """
     Runs in a background thread to listen for messages on a Redis Pub/Sub channel.
@@ -1552,8 +1135,19 @@ def listen_for_index_reloads():
 if __name__ == '__main__':
     # --- Register Blueprints ---
     # We register blueprints here to avoid circular imports for RQ workers.
+    # --- Register Blueprints ---
     from app_chat import chat_bp
+    from app_clustering import clustering_bp
+    from app_analysis import analysis_bp
+    from app_annoy import annoy_bp
+
+    # Only chat gets a prefix
     app.register_blueprint(chat_bp, url_prefix='/chat')
+
+    # Others are registered at root (no prefix)
+    app.register_blueprint(clustering_bp)
+    app.register_blueprint(analysis_bp)
+    app.register_blueprint(annoy_bp)
     
     os.makedirs(TEMP_DIR, exist_ok=True)
     # This block runs only when the script is executed directly (e.g., `python app.py`)
